Break up test_rest.py into separate files
The original test_rest.py will only grow larger as we add more endpoints
with more behaviors, so it should be preemptively broken up into
separate files for different tests.

* Create a `tests/rest` directory to hold all rest tests.
* Create a `tests/rest/conftest.py` to hold all fixtures.
* Create a `tests/rest/rest_utils.py` to hold common utilities.
* Create test files for each tested rest endpoint.

No logic was changed, things were only moved around.
